+++ framework/console/Application.php
@@ -79,14 +79,6 @@ class Application extends \yii\base\Application
      * @var Controller|null the currently active controller instance
      */
     public $controller;
-    /**
-     * Indicates whether jQuery should be used for client-side validation and widgets.
-     *
-     * When enabled, jQuery assets will be registered for validators and widgets that require it.
-     *
-     * @since 22.0
-     */
-    public bool $useJquery = false;
 
     /**
      * {@inheritdoc}