+++ framework/requirements/views/web/index.php
@@ -12,7 +12,7 @@
     <?php $this->renderViewFile(dirname(__FILE__) . '/css.php'); ?>
 </head>
 <body>
-<div class="container">
+<div class="req-container">
     <header>
         <h1>Yii Application Requirement Checker</h1>
     </header>
@@ -34,22 +34,22 @@
 
         <h3>Conclusion</h3>
         <?php if ($summary['errors'] > 0): ?>
-            <div class="alert alert-danger">
+            <div class="req-alert req-alert-danger">
                 <strong>Unfortunately your server configuration does not satisfy the requirements by this application.<br>Please refer to the table below for detailed explanation.</strong>
             </div>
         <?php elseif ($summary['warnings'] > 0): ?>
-            <div class="alert alert-info">
+            <div class="req-alert req-alert-info">
                 <strong>Your server configuration satisfies the minimum requirements by this application.<br>Please pay attention to the warnings listed below and check if your application will use the corresponding features.</strong>
             </div>
         <?php else: ?>
-            <div class="alert alert-success">
+            <div class="req-alert req-alert-success">
                 <strong>Congratulations! Your server configuration satisfies all requirements.</strong>
             </div>
         <?php endif; ?>
 
         <h3>Details</h3>
 
-        <table class="table table-bordered">
+        <table class="req-table req-table-bordered">
             <tr><th>Name</th><th>Result</th><th>Required By</th><th>Memo</th></tr>
             <?php foreach ($requirements as $requirement): ?>
             <tr class="<?php echo $requirement['condition'] ? 'success' : ($requirement['mandatory'] ? 'danger' : 'warning') ?>">