+++ tests/framework/validators/TrimValidatorTest.php
@@ -10,6 +10,7 @@ namespace yiiunit\framework\validators;
 
 use yii\base\DynamicModel;
 use yii\validators\TrimValidator;
+use yii\validators\Validator;
 use yiiunit\TestCase;
 
 /**
@@ -17,10 +18,17 @@ use yiiunit\TestCase;
  */
 class TrimValidatorTest extends TestCase
 {
+    use ClientScriptDispatchTestTrait;
+
     protected function setUp(): void
     {
         parent::setUp();
         $this->destroyApplication();
+    }
+
+    protected function createValidatorInstance(array $config = []): Validator
+    {
+        return new TrimValidator($config);
     }
 
     public function testTrimWhitespace(): void